move to the file automatically generated categorizeGenTtbar_cfi
